Remove OS X Xcode compiler warnings.
git-svn-id: file:///fltk/svn/fltk/branches/branch-1.3@7726 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/Fl_Text_Buffer.cxx
@@ -1176,7 +1176,7 @@ int Fl_Text_Buffer::search_forward(int startPos, const char *searchString,
         return 1;
       }
       // FIXME: character is ucs-4
-    } while ((matchCase ? character(bp++) == *sp++ :
+    } while ((matchCase ? character(bp++) == (unsigned int)*sp++ :
               toupper(character(bp++)) == toupper(*sp++))
              && bp < length());
     startPos++;
@@ -1201,7 +1201,7 @@ int Fl_Text_Buffer::search_backward(int startPos, const char *searchString,
         return 1;
       }
       // FIXME: character is ucs-4
-    } while ((matchCase ? character(bp--) == *sp-- :
+    } while ((matchCase ? character(bp--) == (unsigned int)*sp-- :
               toupper(character(bp--)) == toupper(*sp--))
              && bp >= 0);
     startPos--;
